resources/python_requirements.txt
Django == 1.10.2
psycopg2
gunicorn
django-extensions
ipython
django-model-utils
geopandas
Faker
model_mommy
git+https://github.com/sigma-geosistemas/mommy_spatial_generators.git#egg=mommy_spatial_generators
requests
clint
djangorestframework
markdown
django-filter
radar
ptpython
pygraphviz
